requirements.txt

pillow==10.0.1
requests==2.31.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0